Data service was making multiple requests for the same data. Fixed.
Each tile is a neighbor to 8 other tiles. If a neighbor requested
a tile for data while it has not been loaded, the old code would
create a new request for the data. This happened on all the tiles
since all the requests flood through in a very short interval.