See detailed description
Async is fine w/ modern version of node on dev. The prior commit was ONLY for `uat` which was still behind Node versions. `dev` is up to speed, so reincluding this as a separate commit so as to not create conflicts when merging to `uat` in the future. (The other commits were cherry picked.)